[jsk_perception] Use DiagnosticNodelet to improve subscribing
